+++ src/states/HowToPlayState.cpp
@@ -61,24 +61,24 @@ void HowToPlayState::draw(SDL_Renderer * renderer) {
     loadTexts();
 
     // Draw title
-    SDL_FRect rect_title_text = {this->options->getScreenWidth() / 2, this->options->getScreenHeight() / 8, 0, 0};
+    SDL_FRect rect_title_text = {(float) this->options->getScreenWidth() / 2.0f, (float) this->options->getScreenHeight() / 8.0f, 0.0f, 0.0f};
     SDL_GetTextureSize(this->text_title, &rect_title_text.w, &rect_title_text.h);
-    rect_title_text.x -= rect_title_text.w / 2;
-    rect_title_text.y -= rect_title_text.h / 2;
+    rect_title_text.x -= rect_title_text.w / 2.0f;
+    rect_title_text.y -= rect_title_text.h / 2.0f;
     SDL_RenderTexture(renderer, this->text_title, NULL, &rect_title_text);
 
     // Draw bottom text
-    SDL_FRect rect_bottom_text = {this->options->getScreenWidth() / 2, this->options->getScreenHeight(), 0, 0};
+    SDL_FRect rect_bottom_text = {(float) this->options->getScreenWidth() / 2.0f, (float) this->options->getScreenHeight(), 0.0f, 0.0f};
     SDL_GetTextureSize(this->text_bottom, &rect_bottom_text.w, &rect_bottom_text.h);
-    rect_bottom_text.x -= rect_bottom_text.w / 2;
-    rect_bottom_text.y -= rect_bottom_text.h * 1.5;
+    rect_bottom_text.x -= rect_bottom_text.w / 2.0f;
+    rect_bottom_text.y -= rect_bottom_text.h * 1.5f;
     SDL_RenderTexture(renderer, this->text_bottom, NULL, &rect_bottom_text);
 
     // Draw background
     SDL_FRect rect_background;
-    rect_background.x = 0;
-    rect_background.y = this->options->getScreenHeight()/4;
-    rect_background.w = this->options->getScreenWidth();
+    rect_background.x = 0.0f;
+    rect_background.y = (float) this->options->getScreenHeight() / 4.0f;
+    rect_background.w = (float) this->options->getScreenWidth();
     rect_background.h = rect_bottom_text.y - rect_background.y;
     SDL_SetRenderDrawColor(renderer, COLOR_BOARD.r, COLOR_BOARD.g, COLOR_BOARD.b, COLOR_BOARD.a);
     SDL_RenderFillRect(renderer, &rect_background);
@@ -98,36 +98,36 @@ void HowToPlayState::draw(SDL_Renderer * renderer) {
     );
 
     // Draw everything else
-    SDL_FRect rect_goal = {this->options->getScreenWidth() / 2, this->text_start_y, 0, 0};
+    SDL_FRect rect_goal = {(float) this->options->getScreenWidth() / 2.0f, (float) this->text_start_y, 0.0f, 0.0f};
     SDL_GetTextureSize(this->text_goal, &rect_goal.w, &rect_goal.h);
-    rect_goal.x -= rect_goal.w / 2;
+    rect_goal.x -= rect_goal.w / 2.0f;
     SDL_RenderTexture(renderer, this->text_goal, NULL, &rect_goal);
 
-    SDL_FRect rect_match_shells = {this->options->getScreenWidth() / 2, rect_goal.y + rect_goal.h, 0, 0};
+    SDL_FRect rect_match_shells = {(float) this->options->getScreenWidth() / 2.0f, rect_goal.y + rect_goal.h, 0.0f, 0.0f};
     SDL_GetTextureSize(this->text_match_shells, &rect_match_shells.w, &rect_match_shells.h);
-    rect_match_shells.x -= rect_match_shells.w / 2;
+    rect_match_shells.x -= rect_match_shells.w / 2.0f;
     SDL_RenderTexture(renderer, this->text_match_shells, NULL, &rect_match_shells);
 
-    SDL_FRect rect_shells_src = {0, 0, this->options->getShellSize() * 5, this->options->getShellSize()};
-    SDL_FRect rect_shells_dest = {(int)(this->options->getScreenWidth() / 2) - (int)(this->options->getShellSize() * 2.5), rect_match_shells.y + rect_match_shells.h, rect_shells_src.w, rect_shells_src.h};
+    SDL_FRect rect_shells_src = {0.0f, 0.0f, (float) this->options->getShellSize() * 5.0f, (float) this->options->getShellSize()};
+    SDL_FRect rect_shells_dest = {(float) this->options->getScreenWidth() / 2.0f - (float) this->options->getShellSize() * 2.5f, rect_match_shells.y + rect_match_shells.h, rect_shells_src.w, rect_shells_src.h};
     SDL_RenderTexture(renderer, textures.getShellTexture(), &rect_shells_src, &rect_shells_dest);
 
-    SDL_FRect rect_match_bubbles = {this->options->getScreenWidth() / 2, rect_shells_dest.y + rect_shells_dest.h, 0, 0};
+    SDL_FRect rect_match_bubbles = {(float) this->options->getScreenWidth() / 2.0f, rect_shells_dest.y + rect_shells_dest.h, 0.0f, 0.0f};
     SDL_GetTextureSize(this->text_match_bubbles, &rect_match_bubbles.w, &rect_match_bubbles.h);
-    rect_match_bubbles.x -= rect_match_bubbles.w / 2;
+    rect_match_bubbles.x -= rect_match_bubbles.w / 2.0f;
     SDL_RenderTexture(renderer, this->text_match_bubbles, NULL, &rect_match_bubbles);
 
-    SDL_FRect rect_bubble_src = {this->options->getShellSize() * 6, 0, this->options->getShellSize(), this->options->getShellSize()};
-    SDL_FRect rect_bubble_dest = {this->options->getScreenWidth() / 2 - this->options->getShellSize() / 2, rect_match_bubbles.y + rect_match_bubbles.h, rect_bubble_src.w, rect_bubble_src.h};
+    SDL_FRect rect_bubble_src = {(float) this->options->getShellSize() * 6.0f, 0.0f, (float) this->options->getShellSize(), (float) this->options->getShellSize()};
+    SDL_FRect rect_bubble_dest = {(float) this->options->getScreenWidth() / 2.0f - (float) this->options->getShellSize() / 2.0f, rect_match_bubbles.y + rect_match_bubbles.h, rect_bubble_src.w, rect_bubble_src.h};
     SDL_RenderTexture(renderer, textures.getShellTexture(), &rect_bubble_src, &rect_bubble_dest);
 
-    SDL_FRect rect_match_urchins = {this->options->getScreenWidth() / 2, rect_bubble_dest.y + rect_bubble_dest.h, 0, 0};
+    SDL_FRect rect_match_urchins = {(float) this->options->getScreenWidth() / 2.0f, rect_bubble_dest.y + rect_bubble_dest.h, 0.0f, 0.0f};
     SDL_GetTextureSize(this->text_match_urchins, &rect_match_urchins.w, &rect_match_urchins.h);
     rect_match_urchins.x -= rect_match_urchins.w / 2;
     SDL_RenderTexture(renderer, this->text_match_urchins, NULL, &rect_match_urchins);
 
-    SDL_FRect rect_urchin_src = {this->options->getShellSize() * 5, 0, this->options->getShellSize(), this->options->getShellSize()};
-    SDL_FRect rect_urchin_dest = {this->options->getScreenWidth() / 2 - this->options->getShellSize() / 2, rect_match_urchins.y + rect_match_urchins.h, rect_urchin_src.w, rect_urchin_src.h};
+    SDL_FRect rect_urchin_src = {(float) this->options->getShellSize() * 5.0f, 0.0f, (float) this->options->getShellSize(), (float) this->options->getShellSize()};
+    SDL_FRect rect_urchin_dest = {(float) this->options->getScreenWidth() / 2.0f - (float) this->options->getShellSize() / 2.0f, rect_match_urchins.y + rect_match_urchins.h, rect_urchin_src.w, rect_urchin_src.h};
     if (rect_urchin_dest.y + rect_urchin_dest.h < rect_bottom_text.y) {
         SDL_RenderTexture(renderer, textures.getShellTexture(), &rect_urchin_src, &rect_urchin_dest);
     }
